add check need recovery softcenter

+++ onekey.sh
@@ -3,13 +3,21 @@
 somecheck(){
 	# 检查是否为ks固件
 	if [ ! -d /koolshare/bin/ ]; then
-		echo "🔴 请在koolcenter下执行，其他环境可能存在问题。"
-		exit;
+		echo "🔴 请在koolcenter固件下执行修复"
+		exit
 	fi
 	# 检查固件脚本目录
 	if [ ! -d /rom/etc/koolshare/scripts/ ]; then
-		echo "🔴 固件脚本目录不存在，无法一键修复。"
-		exit;
+		echo "🔴 固件脚本目录不存在，无法修复"
+		exit
+	fi
+	if [ -n $(which dbus) ]; then
+		echo "🔴 未找到dbus命令，无法修复"
+		exit
+	fi
+	if [ -n $(cat /rom/etc/koolshare/.soft_ver) ];then
+		echo "🔴 未找到固件版本号，无法修复"
+		exit
 	fi
 }
 
@@ -17,9 +25,10 @@ changeAsd(){
 	local file="asd"
 	local file_path=$(which "$file")
 	local target_dir="/koolshare/bin/"
+	echo "😛 Step 1: 处理 $file "
 	# 检查文件是否找到
 	if [ -n "$file_path" ]; then
-		echo "✅️ 找到 $file，路径为 $file_path"
+		echo "ℹ️ 找到 $file，路径为 $file_path"
 
 		# 构建目标挂载点
 		mount_point="$target_dir/$file"
@@ -27,7 +36,7 @@ changeAsd(){
 		# 检查是否已fuck
 		checkFuck=$(ps |grep asd |grep /bin/sh)
 		if [ -n "$checkFuck" ];then
-			echo "✅️ 已处理 $file 无需再次处理！"
+			echo "ℹ️ 已处理 $file 无需再次处理"
 			return
 		fi
 
@@ -54,54 +63,61 @@ done" > $mount_point
 		local checkFuck
 		until [ -n "${checkFuck}" ]; do
 			usleep 250000
-			echo "ℹ️ 检查处理状态中..."
+			echo "ℹ️ 检查处理状态中"
 			i=$(($i - 1))
 			checkFuck=$(ps |grep asd |grep /bin/sh)
 			if [ "$i" -lt 1 ]; then
-				echo "🔴 处理失败，请稍后再次尝试。"
-				return
+				echo "🔴 处理失败，请稍后再次尝试"
+				# 处理失败直接结束 恢复了软件中心也会被删
+				exit
 			fi
 		done
 
-		echo "🆗 处理 $file 成功"
+		echo "ℹ️ 处理 $file 成功"
 
-		echo "ℹ️  $file 处理重启失效，如果需要持久化，请前往相应群组了解详情。"
+		echo "ℹ️  $file 处理重启失效，如果需要持久化，请前往相应群组了解详情"
 	else
 		echo "ℹ️  $file 未找到，您不需要处理"
 	fi
+	echo "✅️ Step 1 Done!"
+	echo ""
 }
 
 recoverySoftcenter(){
-	# 恢复文件
-	echo "ℹ️  恢复软件中心版本号"
-	cp -rf /rom/etc/koolshare/.soft_ver /koolshare/ >/dev/null 2>&1
-	cp -rf /rom/etc/koolshare/.soft_ver_old /koolshare/  >/dev/null 2>&1
-	# 写入版本号dbus值
-	dbus set softcenter_version=$(cat /koolshare/.soft_ver)
-	echo "ℹ️  恢复软件中心二进制"
-	cp -rf /rom/etc/koolshare/bin/* /koolshare/bin/  >/dev/null 2>&1
-	echo "ℹ️  恢复软件中心资源"
-	cp -rf /rom/etc/koolshare/res/* /koolshare/res/  >/dev/null 2>&1
-	cp -rf /rom/etc/koolshare/webs/* /koolshare/webs/  >/dev/null 2>&1
-	echo "ℹ️  恢复软件中心脚本"
-	cp -rf /rom/etc/koolshare/scripts/* /koolshare/scripts/  >/dev/null 2>&1
-	cp -rf /rom/etc/koolshare/perp/* /koolshare/perp/  >/dev/null 2>&1
+	echo "😛 Step 2: 恢复软件中心 "
+	if [ ! -f /koolshare/scripts/ks_app_install.sh ] || [ ! -f /koolshare/scripts/ks_tar_install.sh ];then
+		# 恢复文件
+		echo "ℹ️  恢复软件中心版本号"
+		cp -rf /rom/etc/koolshare/.soft_ver /koolshare/ >/dev/null 2>&1
+		cp -rf /rom/etc/koolshare/.soft_ver_old /koolshare/  >/dev/null 2>&1
+		# 写入版本号dbus值
+		dbus set softcenter_version=$(cat /koolshare/.soft_ver)
+		echo "ℹ️  恢复软件中心二进制"
+		cp -rf /rom/etc/koolshare/bin/* /koolshare/bin/  >/dev/null 2>&1
+		echo "ℹ️  恢复软件中心资源"
+		cp -rf /rom/etc/koolshare/res/* /koolshare/res/  >/dev/null 2>&1
+		cp -rf /rom/etc/koolshare/webs/* /koolshare/webs/  >/dev/null 2>&1
+		echo "ℹ️  恢复软件中心脚本"
+		cp -rf /rom/etc/koolshare/scripts/* /koolshare/scripts/  >/dev/null 2>&1
+		cp -rf /rom/etc/koolshare/perp/* /koolshare/perp/  >/dev/null 2>&1
 
-	# 文件赋权
-	chmod +x /koolshare/scripts/*
-	chmod +x /koolshare/perp/perp.sh
-	# 重启软件中心
-	if [ -f /koolshare/perp/perp.sh ];then
-		echo "ℹ️  重启软件中心"
-		sh /koolshare/perp/perp.sh
+		# 文件赋权
+		chmod +x /koolshare/scripts/*  >/dev/null 2>&1
+		chmod +x /koolshare/perp/perp.sh  >/dev/null 2>&1
+		# 重启软件中心
+		if [ -f /koolshare/perp/perp.sh ];then
+			echo "ℹ️  重启软件中心"
+			sh /koolshare/perp/perp.sh
+		fi
+	else
+		echo "ℹ️  软件中心无需恢复"
 	fi
-	echo "✅ 软件中心恢复成功"
+	echo "✅️ Step 2 Done!"
+	echo ""
 }
 
 somecheck
 changeAsd
 recoverySoftcenter
 
-echo "🆗 enjoy ~"
-#空一行，看结果
-echo ""
+echo "🆗 all done, enjoy it~"
